refact(score): refactor score logic
