--- src/main/java/bowling/domain/score/Bonus.java
@@ -1,11 +1,0 @@
-package bowling.domain.score;
-
-public enum Bonus {
-    STRIKE(2), SPARE(1);
-
-    private final int bonus;
-
-    Bonus(int bonus){
-        this.bonus = bonus;
-    }
-}
